[C++] Fix compiler warnings.

+++ aeron-client/src/test/c/concurrent/aeron_distinct_error_log_test.cpp
@@ -340,7 +340,7 @@ static void test_append_distinct_message(aeron_distinct_error_log_t *error_log, 
 
     size_t entries =
         aeron_error_log_read(log_buffer->data(), log_buffer->size(), distinct_message_log_reader, &counts, 0);
-    ASSERT_EQ(entries, NUM_THREADS) << "invalid number of messages";
+    ASSERT_EQ(entries, (size_t)NUM_THREADS) << "invalid number of messages";
 
     for (int count : counts)
     {
@@ -407,8 +407,8 @@ static void test_update_same_message(aeron_distinct_error_log_t *error_log, buff
     size_t count = 0;
     size_t entries =
         aeron_error_log_read(log_buffer->data(), log_buffer->size(), same_message_log_reader, &count, 0);
-    ASSERT_EQ(entries, 1) << "message appended multiple times";
-    ASSERT_EQ(count,  APPENDS_PER_THREAD * NUM_THREADS) << "missing observations";
+    ASSERT_EQ(entries, (size_t)1) << "message appended multiple times";
+    ASSERT_EQ(count,  (size_t)(APPENDS_PER_THREAD * NUM_THREADS)) << "missing observations";
 }
 
 TEST_F(DistinctErrorLogTest, concurrentAppendSameMessage)
